rename testing utilites to match directory name
