+++ pkg/util/testingjobs/pytorchjob/wrappers_pytorchjob.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package testing
+package pytorchjob
 
 import (
 	kftraining "github.com/kubeflow/training-operator/pkg/apis/kubeflow.org/v1"
